+++ packages/interweave/src/index.js
@@ -12,15 +12,18 @@ import EmailMatcher from './matchers/EmailMatcher';
 import HashtagMatcher from './matchers/HashtagMatcher';
 import IpMatcher from './matchers/IpMatcher';
 import UrlMatcher from './matchers/UrlMatcher';
+import { FilterShape, MatcherShape } from './shapes';
 
 export {
   Markup,
   Matcher,
+  MatcherShape,
   EmailMatcher,
   HashtagMatcher,
   IpMatcher,
   UrlMatcher,
   Filter,
+  FilterShape,
 };
 
 export default Interweave;