+++ packages/interweave/src/types.js
@@ -4,7 +4,7 @@
  * @flow
  */
 
-/* eslint-disable */
+/* eslint-disable no-use-before-define, flowtype/sort-keys */
 
 export interface NodeInterface {
   attributes?: NamedNodeMap,
@@ -27,6 +27,25 @@ export type NodeConfig = {
 };
 
 export type Attributes = { [key: string]: string | number | boolean };
+
+export type AfterParseCallback = (content: React$Node[], props: Object) => React$Node[];
+
+export type BeforeParseCallback = (content: string, props: Object) => string;
+
+export interface FilterInterface {
+  attribute: string,
+  filter(value: string): string,
+}
+
+export interface MatcherInterface {
+  inverseName: string,
+  propName: string,
+  asTag(): string,
+  createElement(match: string, props?: Object): React$Node,
+  match(value: string): ?MatchResponse,
+  onBeforeParse(content: string, props: Object): string,
+  onAfterParse(content: React$Node[], props: Object): React$Node[],
+}
 
 export type MatcherFactory = (match: string, props: Object) => React$Node;